feed2social/requirements.txt
feedparser
atproto
Mastodon.py
requests
beautifulsoup4
markdownify
b2sdk